% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_updated_dist}
\alias{plot_updated_dist}
\title{Plot updated distributions}
\usage{
plot_updated_dist(
  updated_dist_params_obj,
  vonmises_mu = NULL,
  include_random_effect = FALSE,
  include_tentative = TRUE
)
}
\arguments{
\item{updated_dist_params_obj}{An updatedDistributionParameters instance created from calling update_dist on an ISSA model}

\item{vonmises_mu}{pi or 0, default NULL. Tells the function where to center the distribution when plotting a von Mises}

\item{include_random_effect}{TRUE or FALSE, default F. Indicates whether or not to plot the curves per random effect variable}

\item{include_tentative_dist}{TRUE or FALSE, default T. Indicates whether or not to plot the tentative distribution}
}
\value{
ggplot object
}
\description{
Plot updated distributions
}
\examples{

library(dplyr)
library(amt)
library(glmmTMB)

model <- glmmTMB(
  case_ ~ sl_ + log_sl_ + sl_:sex + log_sl_:sex,
  data = mixedssa_fisher_data
)

updated_params <- update_dist(model,
  dist_name = "gamma",
  interaction_var_name = "sex",
  beta_sl = "sl_",
  beta_log_sl = "log_sl_"
)

plot_updated_dist(updated_params)


model <- glmmTMB(
  case_ ~ cos_ta_ + cos_ta_:elevation + (0 + cos_ta_ | id),
  data = mixedssa_fisher_data
)

updated_params <- update_dist(model,
  dist_name = "vonmises",
  interaction_var_name = "elevation",
  random_effects_var_name = "id",
  beta_cos_ta = "cos_ta_",
  tentative_dist = tentative_dist
)


plot_updated_dist(updated_params,
  vonmises_mu = 0,
  include_random_effect = T,
  include_tentative = F)

}
