% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_categorical.R
\name{update_dist_by_categorical_var}
\alias{update_dist_by_categorical_var}
\title{Update movement distributions based on fitted models}
\usage{
update_dist_by_categorical_var(
  model,
  dist_name,
  random_effects_var_name,
  interaction_var_name,
  coef_names
)
}
\description{
Update movement distributions based on fitted models
}
