% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_continous.R
\name{update_distributions_by_continuous_var}
\alias{update_distributions_by_continuous_var}
\title{Update movement distributions based on fitted models}
\usage{
update_distributions_by_continuous_var(
  data,
  model,
  dist_name,
  interaction_var_name,
  quantiles = c(0.5, 0.5, 0.95),
  coef_names = NULL
)
}
\description{
Update movement distributions based on fitted models
}
