% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{plot_updated_dist}
\alias{plot_updated_dist}
\title{Plot updated distributions}
\usage{
plot_updated_dist(
  updated_dist_params_obj,
  vonmises_mu = NULL,
  xlim = NULL,
  include_random_effect = FALSE,
  include_tentative = TRUE,
  print_plot = TRUE
)
}
\arguments{
\item{updated_dist_params_obj}{\verb{[updatedDistributionParameters]} The output from calling update_dist on an ISSA model}

\item{vonmises_mu}{\verb{[numeric(1)]\{NULL\}}, pi or 0, optional parameter. Tells the function where to center the distribution when plotting a von Mises}

\item{xlim}{\verb{[numeric(1)]\{NULL\}}, the max step length of the density curve, default is the max step length in the model's data. Ignored if distribution is von Mises}

\item{include_random_effect}{\verb{[logical(1)]\{FALSE\}}. Indicates whether or not to plot the curves per random effect variable}

\item{include_tentative}{\verb{[logical(1)]\{TRUE\}}. Indicates whether or not to plot the tentative distribution}

\item{print_plot}{\verb{[logical(1)]\{TRUE\}}. Indicates whether or not to plot the plot before returning it}
}
\value{
ggplot object
}
\description{
Plot updated distributions
}
\examples{

library(dplyr)
library(amt)
library(glmmTMB)

model <- glmmTMB(
  case_ ~ sl_ + log_sl_ + sl_:sex + log_sl_:sex,
  data = mixedssa_fisher_data
)

updated_params <- update_dist(model,
  dist_name = "gamma",
  interaction_var_name = "sex",
  beta_sl = "sl_",
  beta_log_sl = "log_sl_"
)

plot_updated_dist(updated_params, xlim = 100)


model <- glmmTMB(
  case_ ~ cos_ta_ + cos_ta_:elevation + (0 + cos_ta_ | id),
  data = mixedssa_fisher_data
)


tentative_dist_data <- mixedssa_fisher_data \%>\%
  dplyr::filter(case_ == TRUE) \%>\%
  dplyr::pull("cos_ta_")

tentative_dist <- fit_distr(
  tentative_dist_data,
  dist_name = "vonmises", na.rm = TRUE
)

updated_params <- update_dist(model,
  dist_name = "vonmises",
  interaction_var_name = "elevation",
  random_effects_var_name = "id",
  beta_cos_ta = "cos_ta_",
  tentative_dist = tentative_dist
)


plot_updated_dist(updated_params,
  vonmises_mu = 0,
  include_random_effect = TRUE,
  include_tentative = FALSE
)

}
