% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_update.R
\name{update_dist}
\alias{update_dist}
\title{Update movement distributions based on fitted models}
\usage{
update_dist(
  model,
  dist_name,
  beta_sl = NULL,
  beta_log_sl = NULL,
  beta_sl_sq = NULL,
  beta_log_sl_sq = NULL,
  beta_cos_ta = NULL,
  random_effects_var_name = NULL,
  interaction_var_name = NULL,
  quantiles = DEFAULT_QUANTILES
)
}
\description{
Update movement distributions based on fitted models
}
