% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{update_distributions_by_categorical_var}
\alias{update_distributions_by_categorical_var}
\title{Update movement distributions based on fitted models}
\usage{
update_distributions_by_categorical_var(
  data,
  model,
  dist_name,
  coef_names = NULL,
  reference_category = "reference_category"
)
}
\description{
Update movement distributions based on fitted models
}
